requirements.txt
python-telegram-bot==22.6
requests==2.32.5
